Fix mov insert

+++ src/main/java/h4rk/finance/repository/MovCatRepository.java
@@ -3,7 +3,7 @@ package h4rk.finance.repository;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-
+import java.math.BigInteger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
@@ -14,7 +14,7 @@ public class MovCatRepository {
 	@Autowired
 	private JdbcTemplate jdbcTemplate;
 	
-	public int postMovCat(long movId, List<Long> catIds) {
+	public int postMovCat(BigInteger movId, List<Long> catIds) {
 		int[] i = jdbcTemplate.batchUpdate("INSERT INTO mov_cat (mov_id, cat_id) VALUES (?, ?)", catIds.stream().map(catId -> new Object[] {movId, catId}).collect(Collectors.toList()));
 		if(Arrays.stream(i).sum()==catIds.size()) {
 			return Arrays.stream(i).sum();
