+++ src/main/java/h4rk/finance/service/MovCatService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
+import java.math.BigInteger;
 import h4rk.finance.repository.MovCatRepository;
 
 @Service
@@ -12,7 +12,7 @@ public class MovCatService {
 	@Autowired
 	private MovCatRepository movCatRepository;
 
-	public void postMovCat(long movId, List<Long> catIds) {
+	public void postMovCat(BigInteger movId, List<Long> catIds) {
 		movCatRepository.postMovCat(movId, catIds);
 	}
 }
